requirements.txt
streamlit
fastapi
uvicorn
pydantic
requests
python-dotenv
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
llama-parse
chromadb
tavily-python
tiktoken
google-generativeai
pytest
httpx
pillow>=11.0.0